move all ui/event logic from the to the VM
